refactor(repo): refactor utils to tangle-shared-ui
- refactor `extractErrorFromStatus` to tangle-shared-ui
- refactor `optimizeTxBatch` to tangle-shared-ui
- refactor `useTxNotification` to tangle-shared-ui
