+++ libs/tangle-shared-ui/src/utils/optimizeTxBatch.ts
